backend/requirements.txt
Flask==3.0.3
jsonify==0.5
# psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1